+++ src/pytask_parallel/execute.py
@@ -210,6 +210,14 @@ class ProcessesNameSpace:
         return None
 
 
+def _raise_exception_on_breakpoint(*args: Any, **kwargs: Any) -> None:  # noqa: ARG001
+    raise RuntimeError(
+        "You cannot use 'breakpoint()' while parallelizing the execution of tasks with "
+        "pytask-parallel. Please, remove the breakpoint or run the task without "
+        "parallelization to debug it."
+    )
+
+
 def _execute_task(  # noqa: PLR0913
     task: PTask,
     kwargs: dict[str, Any],
@@ -225,6 +233,12 @@ def _execute_task(  # noqa: PLR0913
 
     """
     __tracebackhide__ = True
+
+    # Path sys.breakpointhook to intercept any call of breakpoint() in a subprocess and
+    # print a better exception message.
+    import sys
+
+    sys.breakpointhook = _raise_exception_on_breakpoint
 
     with warnings.catch_warnings(record=True) as log:
         # mypy can't infer that record=True means log is not None; help it.