+++ CHANGES.md
@@ -12,6 +12,8 @@ releases are available on [PyPI](https://pypi.org/project/pytask-parallel) and
 - {pull}`66` deactivates parallelization for dry-runs.
 - {pull}`67` fixes parallelization with partialed task functions.
 - {pull}`68` updates dependencies and syntaxes.
+- {pull}`69` raises more informative error message when `breakpoint()` was uses when
+  parallelizing with processes or loky.
 
 ## 0.3.1 - 2023-05-27